Clean up some #defines...
git-svn-id: http://svn.osgeo.org/mapguide/trunk/MgDev@2699 a442ad82-b40c-0410-ae44-e6d7bd6e4a36

+++ Server/src/Services/Feature/FeatureNumericFunctions.cpp
@@ -523,7 +523,7 @@ void MgFeatureNumericFunctions::GetQuantileCategories(  VECTOR &values, int numC
 
     // How many go into each full bucket?
 
-    int perBucket = (int)ROUND((double)count/(double)numCats);
+    int perBucket = ROUND((double)count/(double)numCats);
     if (perBucket * numCats > count)
         perBucket--;
